Shave bits

+++ arduino/cms/memory.h
@@ -20,8 +20,8 @@ int inst_lux;
 int inst_uv;
 
 // Testing and misc
-int pos = 0;
+byte pos = 0;
 
 // Timeout values, values incremented to count timeout.
-int door_ajar_wait = 80; // How many cycles to count to before closing the door.
+byte door_ajar_wait = 80; // How many cycles to count to before closing the door.
 byte door_ajar_timeout = 0; // Counts up to 255 for as long as the door is ajar.
